Se arregla que el getNumero pedia pasar de long a int

+++ src/src/Pasaje.java
@@ -1,19 +1,19 @@
 public class Pasaje {
-    private Long numero;
+    private long numero;
     private int asiento;
 
     private venta venta;
     private Viaje viaje;
 
-    public Pasaje(Long numero, int asiento, venta venta, Viaje viaje) {
+    public Pasaje(long numero, int asiento, venta venta, Viaje viaje) {
         this.numero = numero;
         this.asiento = asiento;
         this.venta = venta;
         this.viaje = viaje;
     }
 
-    public Long getNumero() {
-        return numero;
+    public long getNumero() {
+        return (int) numero;
     }
 
     public int getAsiento() {
